attached_assets/requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.28
requests==2.31.0
python-dotenv==1.0.1
Pillow==10.2.0
googlemaps==4.10.0
flask-babel==4.0.0
werkzeug==3.0.1 